requirements.txt
Flask==2.3.3
python-docx==0.8.11
requests==2.31.0
beautifulsoup4==4.12.2
Werkzeug==2.3.7
Jinja2==3.1.2
faker==19.6.2
openai>=1.0.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
supabase==2.0.0
reportlab==4.0.4
gunicorn==21.2.0
